Separate all the things and go nuts
